move AOT.MonoPInvokeCallback to NativeCentralManagerProxy
理由は2つ。

1. SafeHandleの役割を安全な解放という1つに止めるため。
2. AOT.MonoPInvokeCallbackへの依存をSafeHandleから外すため。

後者はDocFxで生成するときにSafeHandleを参照するときにUnityEngineを参照しないようにするために必要
